What was changed in the shape class:
- Removed the private variable for rotation, since a rotation variable
isn't needed. Rotation is handled in methods

- Comments were added to the class doc comments which tell users about
the numbers which represent the different types of shapes
- A getType() accessor method was added to allow users to get the type
of shape
- Made the isLeftBreached and isRightBreached variables initially set to
false

Within the rotateCounterClockwise method (these changes will also be
applied to the rotateClockwise method):
- Added paramters like isBottomBreached to detect if the bottom of the
grid was breached

- moveShapeInBounds moves allows shapes to "bounce" off the left, right
and bottom walls. It detects how far out of bounds the shape is, then
shifts it by that amount into the grid

Next steps:
- Add a way for the block to detect if it's rotating into a shape, and
prevent it from rotating