Changes made to Main.java
- Before these changes, the tetris theme did not repeat, stopping after
one play
- Now, the song repeats indefinitely

- The pause button which was previously a rectangle is now a stackpane
containing the previously used rectangle as well as a pause symbol
- If the pause stackpane is clicked, the music is paused and a pause
screen is displayed. This also stops all gameplay.

- Code for the pause screen mainly contains a VBox in which an image
(pause symbol) and 2 labels were placed. The first label indicates that
the game is paused, and the second one says which key to press to return
to gameplay. The space bar must be pressed to return to the game

- A game over screen was also created. This features another VBox
containing a 3 labels. These labels contain a "Game Over" message, total
points, and instructions on what key to restart to start a new game.

- Furthermore, the "root" of the application is now a stackpane. This
allows pause and game over screens to be added on top of the game screen
when needed.

- showPauseScreen() and hidePauseScreen() add/remove the pause screen
from the stackpane called stkAllScreens (the root of the application)

- the showGameOverScreen() method updates the final score and adds the
game over screen to the stackpane. This screen can be removed and the
game can be restarted if the user presses the "n" key.

** CRITICAL ERROR **
- After n is pressed to get rid of the game over screen, none of the
event handlers trigger when the game is restarted. None of the pressed
keys do anything. This must be fixed, which will likely  be the final
change of this project when implemented.