Changes made to Shape class
Changes specifically made to the methods which rotate shapes:
- The rotateClockwise() and rotateCounterClockwise() methods were
combined into one rotateShape() method. This cuts the amount of code in
half, better for efficiency and memory usage.
- Now, an extra parameter must be passed into the rotateShape() method.
A 0 must be entered for a clockwise rotation while a 1 must be entered
for a counterclockwise rotation. This is the third parameter in the
rotateShape() method. The rotateClockwise() method was commented out
instead of deleted in case information within it is needed for future
bug fixes